more cost registration code, as well as some reformatting

+++ LeaderElection/src/common/CostTracker.java
@@ -10,20 +10,20 @@ public class CostTracker {
 		ELECTION, BROADCAST, QUERY, RESPONSE
 	}
 	
-	ConcurrentHashMap<Stage, ConcurrentHashMap<Integer, Integer>> costs;
+	ConcurrentHashMap<Stage, ConcurrentHashMap<Integer, Double>> costs;
 	
 	public CostTracker(int[] ids) {
 		for(Stage s : Stage.values()) {
-			ConcurrentHashMap<Integer, Integer> stageCosts = new ConcurrentHashMap<Integer, Integer>();
+			ConcurrentHashMap<Integer, Double> stageCosts = new ConcurrentHashMap<Integer, Double>();
 			for(Integer id : ids) {
-				stageCosts.put(id, 0);
+				stageCosts.put(id, 0.);
 			}
 			costs.put(s, stageCosts);
 		}
 	}
 	
-	public void registerCosts(Stage s, Integer processID, Integer cost) {
-		ConcurrentHashMap<Integer, Integer> stageCosts = costs.get(s);
+	public void registerCosts(Stage s, Integer processID, Double cost) {
+		ConcurrentHashMap<Integer, Double> stageCosts = costs.get(s);
 		stageCosts.put(processID, stageCosts.get(processID) + cost);
 		costs.put(s, stageCosts);
 	}
